feat(view): implement directive handlers for dynamic binding
- Add handleStyle directive for dynamic style binding in style.ts
- Add handleText directive for setting text content in text.ts
- Create evaluate function for expression evaluation in evaluate.ts
- Refactor view module by moving mount and template creation logic to mount.ts
- Introduce processElement and processChildren functions for directive processing in process.ts
- Define types for binding context, mount options, and directive handlers in types.ts
- Update index.ts to export new module structure and types
