+++ docs/guide/view.md
@@ -2,7 +2,7 @@
 title: View
 ---
 
-The view module provides declarative DOM bindings similar to Vue/Svelte templates, but without requiring a compiler. Bindings are evaluated at runtime using bQuery's reactive system.
+The view module provides declarative DOM bindings similar to Vue/Svelte templates, but without requiring a compiler. Bindings are evaluated at runtime using bQuery's reactive system. Internally, the view module is now split into focused submodules while the public API remains unchanged.
 
 ```ts
 import { mount } from '@bquery/bquery/view';
