+++ src/view/directives/index.ts
@@ -1,0 +1,11 @@
+export { handleBind } from './bind';
+export { handleClass } from './class';
+export { createForHandler } from './for';
+export { handleHtml } from './html';
+export { handleIf } from './if';
+export { handleModel } from './model';
+export { handleOn } from './on';
+export { handleRef } from './ref';
+export { handleShow } from './show';
+export { handleStyle } from './style';
+export { handleText } from './text';
